Refactorized BACK-END and new Postman Collection

+++ BACK-END/src/main/java/org/kuenteco/backend/service/auth/UserServiceImpl.java
@@ -1,34 +1,31 @@
 package org.kuenteco.backend.service.auth;
 
+import java.io.IOException;
+import java.util.Collections;
+import lombok.RequiredArgsConstructor;
+import lombok.extern.slf4j.Slf4j;
+import org.kuenteco.backend.dto.auth.DeleteUserDTO;
 import org.kuenteco.backend.dto.auth.UserDetailDTO;
 import org.kuenteco.backend.entity.User;
 import org.kuenteco.backend.enums.State;
-import org.kuenteco.backend.mapper.UserDetailMapper;
+import org.kuenteco.backend.mapper.auth.UserDetailMapper;
 import org.kuenteco.backend.repository.master.MasterUserRepository;
 import org.kuenteco.backend.repository.slave.SlaveUserRepository;
-import org.springframework.beans.factory.annotation.Autowired;
+import org.kuenteco.backend.service.image.auth.UserImageService;
 import org.springframework.security.core.authority.SimpleGrantedAuthority;
 import org.springframework.security.core.context.SecurityContextHolder;
 import org.springframework.security.core.userdetails.UserDetails;
 import org.springframework.security.core.userdetails.UsernameNotFoundException;
 import org.springframework.stereotype.Service;
 
-import java.util.Collections;
-
+@Slf4j
 @Service
+@RequiredArgsConstructor
 public class UserServiceImpl implements UserService {
     private final SlaveUserRepository slaveUserRepository;
     private final MasterUserRepository masterUserRepository;
-
-    private UserDetailMapper userDetailMapper;
-
-    @Autowired
-    public UserServiceImpl(SlaveUserRepository slaveUserRepository, MasterUserRepository masterUserRepository,
-            UserDetailMapper userDetailMapper) {
-        this.slaveUserRepository = slaveUserRepository;
-        this.masterUserRepository = masterUserRepository;
-        this.userDetailMapper = userDetailMapper;
-    }
+    private final UserImageService imageService;
+    private final UserDetailMapper userDetailMapper;
 
     @Override
     public UserDetails loadUserByUsername(String nameOrEmail) throws UsernameNotFoundException {
@@ -36,26 +33,34 @@ public class UserServiceImpl implements UserService {
         boolean isEmail = nameOrEmail.contains("@");
 
         if (isEmail) {
-            user = slaveUserRepository.findByEmail(nameOrEmail)
-                    .orElseThrow(() -> new UsernameNotFoundException("Datos Invalidos"));
+            user =
+                    slaveUserRepository
+                            .findByEmail(nameOrEmail)
+                            .orElseThrow(() -> new UsernameNotFoundException("Datos Invalidos"));
         } else {
-            user = slaveUserRepository.findByName(nameOrEmail)
-                    .orElseThrow(() -> new UsernameNotFoundException("Datos Invalidos"));
+            user =
+                    slaveUserRepository
+                            .findByUsername(nameOrEmail)
+                            .orElseThrow(() -> new UsernameNotFoundException("Datos Invalidos"));
         }
-        SimpleGrantedAuthority authority = new SimpleGrantedAuthority(user.getRole().getName().toString());
+        SimpleGrantedAuthority authority =
+                new SimpleGrantedAuthority(user.getRole().getName().toString());
 
-        return new org.springframework.security.core.userdetails.User(user.getEmail(), user.getPassword(),
-                Collections.singleton(authority));
+        return new org.springframework.security.core.userdetails.User(
+                user.getEmail(), user.getPassword(), Collections.singleton(authority));
     }
 
     @Override
     public UserDetails loadUserByEmail(String email) throws UsernameNotFoundException {
-        User user = slaveUserRepository.findByEmail(email)
-                .orElseThrow(() -> new UsernameNotFoundException("Datos Invalidos"));
-        SimpleGrantedAuthority authority = new SimpleGrantedAuthority(user.getRole().getName().toString());
+        User user =
+                slaveUserRepository
+                        .findByEmail(email)
+                        .orElseThrow(() -> new UsernameNotFoundException("Datos Invalidos"));
+        SimpleGrantedAuthority authority =
+                new SimpleGrantedAuthority(user.getRole().getName().toString());
 
-        return new org.springframework.security.core.userdetails.User(user.getEmail(), user.getPassword(),
-                Collections.singleton(authority));
+        return new org.springframework.security.core.userdetails.User(
+                user.getEmail(), user.getPassword(), Collections.singleton(authority));
     }
 
     @Override
@@ -63,17 +68,19 @@ public class UserServiceImpl implements UserService {
         boolean isEmail = nameOrEmail.contains("@");
 
         if (isEmail) {
-            return slaveUserRepository.findByEmail(nameOrEmail)
+            return slaveUserRepository
+                    .findByEmail(nameOrEmail)
                     .orElseThrow(() -> new UsernameNotFoundException("Datos Invalidos"));
         } else {
-            return slaveUserRepository.findByName(nameOrEmail)
+            return slaveUserRepository
+                    .findByUsername(nameOrEmail)
                     .orElseThrow(() -> new UsernameNotFoundException("Datos Invalidos"));
         }
     }
 
     @Override
-    public boolean existsByUserName(String name) {
-        return slaveUserRepository.existsByName(name);
+    public boolean existsByUserName(String username) {
+        return slaveUserRepository.existsByUsername(username);
     }
 
     @Override
@@ -89,8 +96,7 @@ public class UserServiceImpl implements UserService {
     @Override
     public void deletePendingEmail(String email) {
         User user = new User();
-        if (user.getAccountState() == State.PENDING)
-            masterUserRepository.removeUserByEmail(email);
+        if (user.getState() == State.PENDING) masterUserRepository.removeUserByEmail(email);
     }
 
     public User getUserDetails() {
@@ -105,7 +111,17 @@ public class UserServiceImpl implements UserService {
         return userDetailMapper.toDto(user);
     }
 
-    public void deteleUser(User masterUser) {
-        masterUserRepository.deleteById(masterUser.getId());
+    @Override
+    public void deteleUser(DeleteUserDTO deleteUserDTOid) throws IOException {
+        User user =
+                slaveUserRepository
+                        .findById(deleteUserDTOid.getId())
+                        .orElseThrow(() -> new UsernameNotFoundException("Usuario no encontrado"));
+
+        if (user.getImage() != null) {
+            imageService.removeImage(user.getImage());
+        }
+
+        masterUserRepository.delete(user);
     }
 }
